+++ src/OrdersList/OrdersListFiltersContainer.js
@@ -3,14 +3,18 @@ import PropTypes from 'prop-types';
 
 import { stripesConnect } from '@folio/stripes/core';
 
+import { getAddresses } from '../common/utils';
+import { ADDRESSES } from '../components/Utils/resources';
 import { reasonsForClosureResource } from '../common/resources';
 import OrdersListFilters from './OrdersListFilters';
 
 const OrdersListFiltersContainer = ({ resources, activeFilters, applyFilters, disabled }) => {
   const closingReasons = resources?.closingReasons?.records;
+  const addresses = getAddresses(resources?.addresses?.records);
 
   return (
     <OrdersListFilters
+      addresses={addresses}
       activeFilters={activeFilters}
       applyFilters={applyFilters}
       closingReasons={closingReasons}
@@ -20,6 +24,7 @@ const OrdersListFiltersContainer = ({ resources, activeFilters, applyFilters, di
 };
 
 OrdersListFiltersContainer.manifest = Object.freeze({
+  addresses: ADDRESSES,
   closingReasons: reasonsForClosureResource,
 });